# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/verifications/merchants": {
      "get": {
        "operationId": "AdminVerificationController_listMerchants",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by profile verification status",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "APPROVED",
                "REJECTED",
                "SUSPENDED"
              ]
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Search query for phone number or display name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-based)",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantVerificationListResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List merchant verification profiles with filtering and search",
        "tags": [
          "Admin Verification"
        ]
      }
    },
    "/api/v1/admin/verifications/merchants/{profileId}": {
      "get": {
        "operationId": "AdminVerificationController_getMerchantDetail",
        "parameters": [
          {
            "name": "profileId",
            "required": true,
            "in": "path",
            "description": "Merchant profile UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantVerificationDetailResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Merchant profile not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get merchant verification profile detail with audit history",
        "tags": [
          "Admin Verification"
        ]
      }
    },
    "/api/v1/admin/verifications/merchants/{profileId}/reveal-nik": {
      "get": {
        "operationId": "AdminVerificationController_revealMerchantNik",
        "parameters": [
          {
            "name": "profileId",
            "required": true,
            "in": "path",
            "description": "Merchant profile UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevealNikResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Merchant profile not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reveal full unmasked NIK for merchant verification",
        "tags": [
          "Admin Verification"
        ]
      }
    },
    "/api/v1/admin/verifications/merchants/{profileId}/documents/ktp": {
      "get": {
        "operationId": "AdminVerificationController_streamMerchantKtp",
        "parameters": [
          {
            "name": "profileId",
            "required": true,
            "in": "path",
            "description": "Merchant profile UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Binary stream of KTP document"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Merchant profile or document not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Stream merchant KTP document",
        "tags": [
          "Admin Verification"
        ]
      }
    },
    "/api/v1/admin/verifications/merchants/{profileId}/approve": {
      "post": {
        "operationId": "AdminVerificationController_approveMerchant",
        "parameters": [
          {
            "name": "profileId",
            "required": true,
            "in": "path",
            "description": "Merchant profile UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveMerchantVerificationActionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantVerificationDetailResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Merchant profile not found"
          },
          "409": {
            "description": "Invalid transition or idempotency conflict"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Approve a pending merchant profile",
        "tags": [
          "Admin Verification"
        ]
      }
    },
    "/api/v1/admin/verifications/merchants/{profileId}/reject": {
      "post": {
        "operationId": "AdminVerificationController_rejectMerchant",
        "parameters": [
          {
            "name": "profileId",
            "required": true,
            "in": "path",
            "description": "Merchant profile UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectMerchantVerificationActionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantVerificationDetailResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or reason missing"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Merchant profile not found"
          },
          "409": {
            "description": "Invalid transition or idempotency conflict"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reject a pending merchant profile with reason",
        "tags": [
          "Admin Verification"
        ]
      }
    },
    "/api/v1/admin/verifications/merchants/{profileId}/suspend": {
      "post": {
        "operationId": "AdminVerificationController_suspendMerchant",
        "parameters": [
          {
            "name": "profileId",
            "required": true,
            "in": "path",
            "description": "Merchant profile UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReasonRequiredVerificationActionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantVerificationDetailResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or reason missing"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Merchant profile not found"
          },
          "409": {
            "description": "Invalid transition or idempotency conflict"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Suspend an approved merchant profile with reason",
        "tags": [
          "Admin Verification"
        ]
      }
    },
    "/api/v1/admin/verifications/merchants/{profileId}/reactivate": {
      "post": {
        "operationId": "AdminVerificationController_reactivateMerchant",
        "parameters": [
          {
            "name": "profileId",
            "required": true,
            "in": "path",
            "description": "Merchant profile UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReasonRequiredVerificationActionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantVerificationDetailResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or reason missing"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Merchant profile not found"
          },
          "409": {
            "description": "Invalid transition or idempotency conflict"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reactivate a suspended merchant profile with reason",
        "tags": [
          "Admin Verification"
        ]
      }
    },
    "/api/v1/admin/verifications/drivers": {
      "get": {
        "operationId": "AdminVerificationController_listDrivers",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by profile verification status",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "APPROVED",
                "REJECTED",
                "SUSPENDED"
              ]
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Search query for phone number or display name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-based)",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverVerificationListResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List driver verification profiles with filtering and search",
        "tags": [
          "Admin Verification"
        ]
      }
    },
    "/api/v1/admin/verifications/drivers/{profileId}": {
      "get": {
        "operationId": "AdminVerificationController_getDriverDetail",
        "parameters": [
          {
            "name": "profileId",
            "required": true,
            "in": "path",
            "description": "Driver profile UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverVerificationDetailResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Driver profile not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get driver verification profile detail with audit history",
        "tags": [
          "Admin Verification"
        ]
      }
    },
    "/api/v1/admin/verifications/drivers/{profileId}/approve": {
      "post": {
        "operationId": "AdminVerificationController_approveDriver",
        "parameters": [
          {
            "name": "profileId",
            "required": true,
            "in": "path",
            "description": "Driver profile UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveVerificationActionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverVerificationDetailResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Driver profile not found"
          },
          "409": {
            "description": "Invalid transition or idempotency conflict"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Approve a pending driver profile",
        "tags": [
          "Admin Verification"
        ]
      }
    },
...
  },
  "info": {
    "title": "TorangGo API",
    "description": "TorangGo backend API foundation. Standard headers: X-Request-ID (request correlation), Idempotency-Key (mutation deduplication), X-Idempotency-Replayed (replayed response indicator).",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "MerchantVerificationListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MerchantVerificationItemDto"
            }
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 20
          },
          "total": {
            "type": "number",
            "example": 42
          }
        },
        "required": [
          "items",
          "page",
          "limit",
          "total"
        ]
      },
...
      "RevealNikResponseDto": {
        "type": "object",
        "properties": {
          "nik": {
            "type": "string",
            "example": "7171012345678901",
            "description": "Full 16-digit unmasked NIK"
          }
        },
        "required": [
          "nik"
        ]
      },
      "ApproveMerchantVerificationActionDto": {
        "type": "object",
        "properties": {
          "expectedSubmissionId": {
            "type": "string",
            "description": "Target submission UUID expected by admin to prevent stale approvals",
            "example": "018f6c5e-8b1b-7a6c-9c3f-4e5f6a7b8c9d"
          },
          "reason": {
            "type": "string",
            "description": "Optional administrative note or approval reason (max 1000 characters)",
            "maxLength": 1000,
            "example": "Identitas dan profil valid",
            "nullable": true
          }
        }
      },
      "RejectMerchantVerificationActionDto": {
        "type": "object",
        "properties": {
          "expectedSubmissionId": {
            "type": "string",
            "description": "Target submission UUID expected by admin to prevent stale rejections",
            "example": "018f6c5e-8b1b-7a6c-9c3f-4e5f6a7b8c9d"
          },
          "reason": {
            "type": "string",
            "description": "Mandatory administrative justification reason (min 3, max 1000 characters)",
            "minLength": 3,
            "maxLength": 1000,
            "example": "Foto KTP buram dan tidak terbaca"
          }
        },
        "required": [
          "reason"
        ]
      },
...
    }
  }
}